fix: Renamed components to match Fish Tank naming conventions and future css module refactor
