+++ src/components/FishTankSwitch.vue
