+++ src/index.ts
@@ -16,9 +16,9 @@ export { default as ButtonOutlineDark} from "./components/ButtonOutlineDark.vue"
 export { default as ButtonOutlineSmall} from "./components/ButtonOutlineSmall.vue"
 export { default as ButtonOutlineDarkSmall} from "./components/ButtonOutlineDarkSmall.vue"
 export { default as Card } from "./components/Card.vue"
-export { default as InputCheckbox } from "./components/InputCheckbox.vue"
-export { default as InputRadio } from "./components/InputRadio.vue"
-export { default as InputSwitch } from "./components/InputSwitch.vue"
+export { default as InputCheckbox } from "./components/FishTankCheckbox.vue"
+export { default as InputRadio } from "./components/FishTankRadio.vue"
+export { default as InputSwitch } from "./components/FishTankSwitch.vue"
 export { default as InputText } from "./components/InputText.vue"
 export { default as Modal } from "./components/Modal.vue"
 export { default as DialogBox } from "./components/DialogBox.vue"
@@ -38,9 +38,9 @@ export { default as FishTankButtonWhite } from "./components/ButtonWhite.vue"
 export { default as FishTankButtonOutline} from "./components/ButtonOutline.vue"
 export { default as FishTankButtonGradient } from "./components/ButtonGradient.vue"
 export { default as FishTankCard } from "./components/Card.vue"
-export { default as FishTankCheckbox } from "./components/InputCheckbox.vue"
-export { default as FishTankRadio } from "./components/InputRadio.vue"
-export { default as FishTankSwitch } from "./components/InputSwitch.vue"
+export { default as FishTankCheckbox } from "./components/FishTankCheckbox.vue"
+export { default as FishTankRadio } from "./components/FishTankRadio.vue"
+export { default as FishTankSwitch } from "./components/FishTankSwitch.vue"
 export { default as FishTankTextInput } from "./components/InputText.vue"
 export { default as FishTankModal } from "./components/Modal.vue"
 export { default as FishTankDialogBox } from "./components/DialogBox.vue"
